merge

+++ icl_memorization/main.py
@@ -5,7 +5,7 @@
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.16.6
+#       jupytext_version: 1.16.2
 #   kernelspec:
 #     display_name: Python (fmri)
 #     language: python
@@ -301,9 +301,9 @@ elif args.wandb_group_name == "memo_may26_zipf_onelayerattention_lr_1e-4_vary_nu
 elif args.wandb_group_name == "memo_nov10_zipf_gpt2_vary_num_hidden_features_num_heads_noresample":
     args.arch = "gpt"
     num_heads = [1] + list(range(2, 17, 2)) # len: 9
-    num_hidden_features = 2 ** np.linspace(0, 9, 10)
+    num_hidden_features = 2 ** np.linspace(0, 9, 10).astype(int)
     args.num_heads = num_heads[args.SLURM_ARRAY_TASK_ID % len(num_heads)] 
-    args.num_hidden_features = num_hidden_features[args.SLURM_ARRAY_TASK_ID // len(num_heads)]
+    args.num_hidden_features = int(num_hidden_features[args.SLURM_ARRAY_TASK_ID // len(num_heads)])
     args.num_layers = 12
     print("SLURM_ARRAY_TASK_ID",args.SLURM_ARRAY_TASK_ID, "num_heads", args.num_heads, "num_layers", args.num_layers)
     
@@ -312,9 +312,9 @@ elif args.wandb_group_name == "memo_nov10_zipf_gpt2_vary_num_hidden_features_num
 elif args.wandb_group_name == "memo_nov10_zipf_gpt2_vary_num_hidden_features_num_heads_doresample":
     args.arch = "gpt"
     num_heads = [1] + list(range(2, 17, 2)) # len: 9
-    num_hidden_features = 2 ** np.linspace(0, 9, 10)
+    num_hidden_features = 2 ** np.linspace(0, 9, 10).astype(int)
     args.num_heads = num_heads[args.SLURM_ARRAY_TASK_ID % len(num_heads)] 
-    args.num_hidden_features = num_hidden_features[args.SLURM_ARRAY_TASK_ID // len(num_heads)]
+    args.num_hidden_features = int(num_hidden_features[args.SLURM_ARRAY_TASK_ID // len(num_heads)])
     args.num_layers = 12
     print("SLURM_ARRAY_TASK_ID",args.SLURM_ARRAY_TASK_ID, "num_heads", args.num_heads, "num_layers", args.num_layers)
